requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-pptx==1.0.2
reportlab==4.2.2
requests==2.32.3
urllib3>=1.21.1,<2
chardet>=3.0.2,<6
python-multipart==0.0.9
pydantic==2.7.1
python-dotenv==1.2.2
